# openapi.yaml
openapi: 3.1.0
info:
  title: PEG HR LMS v1 API
  version: "1.0.0"
  description: >
    HTTP-контракт системы онбординга новых сотрудников (см. SPEC.md).
    Один сервис. Все пути под /api/v1. Реализовано в prototype/.
    Ошибка любого запроса: { error: { code, message, details? }, request_id }.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Err:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "snake_case машинный код" }
            message: { type: string }
            details: {}
    Stage:
      type: string
      enum: [intern, onboarding, completed, archived]
...
    EmployeeRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        position: { type: string }
        position_id: { type: string, format: uuid }
        phone: { type: string }
        start_date: { type: string, format: date }
        stage: { $ref: '#/components/schemas/Stage' }
        onboarding_due_date: { type: [string, "null"], format: date }
        overdue: { type: boolean }
        progress:
          type: object
          properties:
            passed: { type: integer }
            total: { type: integer }
...
security:
  - bearer: []
paths:
...
  /employees:
    get:
      tags: [employees]
      summary: Список сотрудников
      parameters:
        - { name: stage, in: query, schema: { $ref: '#/components/schemas/Stage' } }
        - { name: position_id, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: include_archived, in: query, schema: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/EmployeeRow' } }
    post:
      tags: [employees]
      summary: Завести сотрудника (создаёт User + Employee, stage = intern)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, position_id, phone, start_date, login, password]
              properties:
                full_name: { type: string, maxLength: 160 }
                position_id: { type: string }
                phone: { type: string }
                start_date: { type: string, format: date }
                login: { type: string }
                password: { type: string }
      responses:
        "201":
          description: "{ employee, warnings: string[] } — warnings может содержать no_active_trajectory"
        "422": { $ref: '#/components/responses/Err', description: "phone_exists / login_exists / start_date_too_old" }
  /employees/{id}:
    get:
      tags: [employees]
      summary: Карточка сотрудника (этап, прогресс по блокам/урокам, логин)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: подробная карточка }
        "404": { $ref: '#/components/responses/Err' }
    patch:
      tags: [employees]
      summary: Изменить ФИО / телефон
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                phone: { type: string }
      responses:
        "200": { description: "{ ok: true }" }
        "422": { $ref: '#/components/responses/Err' }
  /employees/{id}/internship-passed:
    post:
      tags: [employees]
      summary: Отметить стажировку пройденной (пробует открыть онбординг)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ ok, onboarding_opened, reason }" }
        "409": { $ref: '#/components/responses/Err', description: "не intern" }
  /employees/{id}/internship-failed:
    post:
      tags: [employees]
      summary: Стажировку не прошёл — удаляет сотрудника и весь прогресс
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ ok, deleted: true }" }
        "409": { $ref: '#/components/responses/Err' }
  /employees/{id}/archive:
    post:
      tags: [employees]
      summary: В архив (user.is_active = false, данные сохраняются)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "{ ok: true }" } }
  /employees/{id}/reset-password:
    post:
      tags: [employees]
      summary: Сбросить пароль — возвращает новый
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "{ password: string }" } }
  /employees/{id}/audit:
    get:
      tags: [employees]
      summary: Журнал действий по сотруднику
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "{ items: [{ ts, actor_login, action, detail }] }" } }
  /employees/{id}/invite:
    post:
      tags: [employees]
      summary: Ссылка-приглашение
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "{ invite_url, login }" } }

  # ---------------- catalog / конструктор (HR) ----------------
...
  /dev/clock:
    get:
      tags: [dev]
      security: []
      summary: "Демо-дата"
      responses: { "200": { description: "{ today, overridden }" } }
    post:
      tags: [dev]
      summary: Установить демо-дату
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [date], properties: { date: { type: string, format: date } } } } } }
      responses: { "200": { description: "{ today, overridden }" } }
  /dev/clock/advance:
    post:
      tags: [dev]
      summary: Сдвинуть демо-дату на N дней
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [days], properties: { days: { type: integer } } } } } }
      responses: { "200": { description: ok } }
...